% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{add_id_date_repeated}
\alias{add_id_date_repeated}
\title{Returns a dataframe with the Id, Date, and Repeated column added, which
are derived from the Animal.ID column.}
\usage{
add_id_date_repeated(df)
}
\arguments{
\item{df}{dataframe having at least the Animal.ID column}
}
\value{
a dataframe with the Id, Date, and Repeated column added
}
\description{
Returns a dataframe with the Id, Date, and Repeated column added, which
are derived from the Animal.ID column.
}
