% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{add_sqlmed_codes}
\alias{add_sqlmed_codes}
\title{Returns df dataframe with the SqlMed procedure_name, procedure_id,
test_id, and test_name values added to the df.}
\usage{
add_sqlmed_codes(df, lab)
}
\arguments{
\item{df}{dataframe with the data from the PDL reports.}

\item{lab}{character vector of length one that indicates the laboratory}
}
\description{
Returns df dataframe with the SqlMed procedure_name, procedure_id,
test_id, and test_name values added to the df.
}
