% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{apply_divisor}
\alias{apply_divisor}
\title{Returns mfi_df with each value divided by the appropriate divisors}
\usage{
apply_divisor(mfi_df, divisors)
}
\arguments{
\item{mfi_df}{dataframe with substracted background}

\item{divisors}{named numeric vector of divisors}
}
\description{
Returns mfi_df with each value divided by the appropriate divisors
}
