% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blank_fill_ids.R
\name{blank_fill_ids}
\alias{blank_fill_ids}
\title{Returns the Ids in right blank filled and upper case animal Ids.}
\usage{
blank_fill_ids(ids, upper = TRUE)
}
\arguments{
\item{ids}{character vector of animal Ids that may need to be transformed
into the proper Id format.}

\item{upper}{logical indicator of whether ids are to be forced to upper case}
}
\description{
Returns the Ids in right blank filled and upper case animal Ids.
}
\examples{
new_ids <- blank_fill_ids(c("12345", "1X1234", "1234","2 3456"))
}
