% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{combine_lines}
\alias{combine_lines}
\title{Returns character vector of length 1 made up of text found in two character
vector elements as subsetted according to start and end locations.}
\usage{
combine_lines(line_1, line_2, start_end)
}
\arguments{
\item{line_1}{first line of text to subset}

\item{line_2}{second line of text to subset}

\item{start_end}{column boundaries to include}
}
\description{
Returns character vector of length 1 made up of text found in two character
vector elements as subsetted according to start and end locations.
}
