% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{create_luminex_screen_results_tbl}
\alias{create_luminex_screen_results_tbl}
\title{Returns TRUE if the creation of the pdl_result table was successful.}
\usage{
create_luminex_screen_results_tbl(conn)
}
\arguments{
\item{conn}{database connection object}
}
\description{
Has the side effect of creating an SQL Server database table to hold the
results received from PDL.
}
