% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{create_spf_samples_tbl}
\alias{create_spf_samples_tbl}
\title{Returns TRUE if the creation of the spf_samples table was successful.}
\usage{
create_spf_samples_tbl(conn)
}
\arguments{
\item{conn}{database connection object}
}
\description{
Has the side effect of creating an SQL Server database table to hold the
list of SPF samples.
}
