% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{divisors_match}
\alias{divisors_match}
\title{Returns a logical value indicating whether or not the list of divisor
names in
the file ``cutoff.xlsx'' is the same as the list within the .lxd file.}
\usage{
divisors_match(file)
}
\arguments{
\item{file}{character vector with the name of the name, including path,
of the XML raw datafile}
}
\description{
Returns a logical value indicating whether or not the list of divisor
names in
the file ``cutoff.xlsx'' is the same as the list within the .lxd file.
}
