% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_pdl_csv_columns.R
\name{fix_pdl_csv_columns}
\alias{fix_pdl_csv_columns}
\title{Returns pdl_df with columns renamed, added and ordered to prepare the
dataframe for inserting into the pdl_result table.}
\usage{
fix_pdl_csv_columns(pdl_df)
}
\arguments{
\item{pdl_df}{dataframe as it comes from read.csv()}
}
\description{
Returns pdl_df with columns renamed, added and ordered to prepare the
dataframe for inserting into the pdl_result table.
}
