% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_luminex_results.R
\name{fmt_luminex_results}
\alias{fmt_luminex_results}
\title{Conditionally formats a worksheet based on data within the dataframe as a
side effect.}
\usage{
fmt_luminex_results(
  excel_file,
  df_list,
  fmtIndexNum,
  wrkSheetIndex,
  sheet_name = "final_result",
  low_positive_controls_df
)
}
\arguments{
\item{excel_file}{character vector with the name of the Excel file}

\item{df_list}{List of dataframes used as the source data for formating and
for writing the Excel file worksheet. The two (data controlling formating
and data written)  may not be the same.}

\item{fmtIndexNum}{Index into the list of dataframes (\code{df_list} used to
guide format highlighting.}

\item{wrkSheetIndex}{Index into the list of dataframes (\code{df_list} used
to write the data into the worksheet.}

\item{sheet_name}{name of the worksheet to be written and highlighted.}

\item{low_positive_controls_df}{dataframe with the low positive controls}
}
\value{
Excel file name of file that has been formatted to highlight
non-negative results in the sheet indicated. This highlights "I" and
"P" results
}
\description{
This function is used for it's side effects. It takes an Excel workbook
(\code{excel_file}), a list of dataframes (\code{df_list}),
an index (\code{fmtIndexNum}) to the dataframe list to
indicate the dataframe to be used for conditional formating, an index
(\code{wrkSheetIndex}) to the dataframe list to indicate the dataframe to
be written into the workbook and then the formatting will be applied to it,
the worksheet name (\code{sheet_name}) of the worksheet being written, and
the dataframe containing values for the low positive controls
(\code{low_positive_controles_df})
}
