% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{format_luminex_results}
\alias{format_luminex_results}
\title{Returns Excel file name of file that has been formatted to highlight
non-negative results in the sheet indicated.}
\usage{
format_luminex_results(excel_file, .df, sheet_name = "final_result",
  low_positive_controls_df)
}
\arguments{
\item{excel_file}{character vector with the name of the Excel file}

\item{.df}{dataframe being used to guide highlighting. It is the same
one as was used to make the worksheet.}

\item{sheet_name}{name of the worksheet to highlight.}

\item{low_positive_controls_df}{dataframe with the low positive controls}
}
\description{
This highlights "I" and "P" results
}
