% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_agents_and_antigen_pairs.R
\name{get_agents_and_antigen_pairs}
\alias{get_agents_and_antigen_pairs}
\title{Returns dataframe of possible antigen pairs and agents}
\usage{
get_agents_and_antigen_pairs()
}
\description{
Returns dataframe of possible antigen pairs and agents
}
