% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_combined.R
\name{get_all_combined}
\alias{get_all_combined}
\title{Returns dataframe with multiple combined results from antigen results}
\usage{
get_all_combined(conn, all_r_mfi_df)
}
\arguments{
\item{conn}{database connection object}

\item{all_r_mfi_df}{antigen results}
}
\description{
The following results set are possible:
 antigen 1 P and antigen 2 P, combined is P
 antigen 1 P and antigen 2 N, combined is I
 antigen 1 N and antigen 2 P, combined in I
 antigen 1 N and antigen 2 N, combined in N
}
