% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{get_all_df}
\alias{get_all_df}
\title{Returns a dataframe with all bead or all summary results combined
depending on which function and column names are passed in.}
\usage{
get_all_df(
  path = "../inst/extdata/",
  .fun = get_summary_df,
  col_names = get_col_names("syntactic_summary")
)
}
\arguments{
\item{path}{file path to directory containing files to be read.}

\item{.fun}{the function object to be used to collect combined results.}

\item{col_names}{character vector of column names in the order they are to
appear.}
}
\description{
Columns missing in some results set are added with NA values.
}
