% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{get_all_files}
\alias{get_all_files}
\title{Returns the list of pathology report files with an extension of ".pdf"
found in folders indicated by the years between start_year and end_year
inclusive.}
\usage{
get_all_files(start_year, end_year)
}
\arguments{
\item{start_year}{integer value indicating the first calendar year as used
in the name of the folders containing reports from the years indicated.}

\item{end_year}{integer value indicating the last calendar year as used in
the name of the folders containing reports from the years indicated.}
}
\description{
Returns the list of pathology report files with an extension of ".pdf"
found in folders indicated by the years between start_year and end_year
inclusive.
}
