% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_regions.R
\name{get_all_regions}
\alias{get_all_regions}
\title{Returns a dataframe with the files, order, and Region[@name]s
for a specific Setup, for example 'DefaultSetup'.}
\usage{
get_all_regions(files, setup_name = "DefaultSetup")
}
\arguments{
\item{files}{character vector of files to be processed}

\item{setup_name}{name of setup in XML file to get regions from}
}
\description{
Returns a dataframe with the files, order, and Region[@name]s
for a specific Setup, for example 'DefaultSetup'.
}
