% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_summary_df.R
\name{get_all_summary_df}
\alias{get_all_summary_df}
\title{Returns a dataframe with all summary results combined.}
\usage{
get_all_summary_df(
  path = "../inst/extdata/",
  col_names = c("Filename", "Animal.ID", "STLV", "SRV", "BV", "SIV", "Measles")
)
}
\arguments{
\item{path}{file path to directory containing files to be read.}

\item{col_names}{character vector of column names in the order they are to
appear.}
}
\description{
Columns missing in some results set are added with NA values.
}
