% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_antigen_pairs.R
\name{get_antigen_pairs}
\alias{get_antigen_pairs}
\title{Returns a dataframe with the agent and two antigens used to assay for it.}
\usage{
get_antigen_pairs(antigens)
}
\arguments{
\item{antigens}{vector of antigens}
}
\description{
Since only one antigen is used for Measles and T_cruzi, they are repeated.
}
\details{
Charles River changed one of their beads without any warning.
Specifically, they dropped SIV gp130 (region 34) and replaced it for
SIV gp120 (region 63).
}
