% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_assay_pairs.R
\name{get_assay_pairs}
\alias{get_assay_pairs}
\title{Returns a dataframe, with sets of assays with an antibody assay (ab_assay)
and a PCR based assay (pcr_assay) for each agent.}
\usage{
get_assay_pairs()
}
\description{
Returns a dataframe, with sets of assays with an antibody assay (ab_assay)
and a PCR based assay (pcr_assay) for each agent.
}
