% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bad_date_ids_and_id_ids.R
\name{get_bad_date_ids_and_id_ids}
\alias{get_bad_date_ids_and_id_ids}
\title{Returns a list with two character vectors containing the items in the
vector dates that are not Ids and the items in ids that are not Ids.}
\usage{
get_bad_date_ids_and_id_ids(conn, ids, dates)
}
\arguments{
\item{conn}{database connection object}

\item{ids}{character vector of possible Ids (or dates)}

\item{dates}{character vector of possible Ids (or dates)}
}
\description{
Returns a list with two character vectors containing the items in the
vector dates that are not Ids and the items in ids that are not Ids.
}
