% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bead_df.R
\name{get_bead_df}
\alias{get_bead_df}
\title{Get bead call multipliers from results section of Excel SPF Report file.}
\usage{
get_bead_df(results)
}
\arguments{
\item{results}{character vector with line from the Excel file worksheet
having only the result table records}
}
\value{
dataframe with bead call results from the results section of the
Excel file.
}
\description{
Takes a dataframe with the results section of the data in an Excel
SPF Report file, reads in the cuttoff values from the sheet, and creates
a dataframe with the bead call results to use as multipliers.
}
