% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bead_headers.R
\name{get_bead_headers}
\alias{get_bead_headers}
\title{Returns a dataframe with all headers found in the bead results as seen by
get_bead_df()}
\usage{
get_bead_headers(path, files)
}
\arguments{
\item{path}{file path to directory containing files to be read.}

\item{files}{names of files to be read}
}
\description{
Returns a dataframe with all headers found in the bead results as seen by
get_bead_df()
}
