% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_col_boundaries.R
\name{get_col_boundaries}
\alias{get_col_boundaries}
\title{Returns an integer vector with the the positions in the line that begin
each column label.}
\usage{
get_col_boundaries(content)
}
\arguments{
\item{content}{charcter vector containing results. First line should contain
column labels.}
}
\description{
Returns an integer vector with the the positions in the line that begin
each column label.
}
