% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_col_names.R
\name{get_col_names}
\alias{get_col_names}
\title{Returns character vector of column names base on type requested.}
\usage{
get_col_names(type)
}
\arguments{
\item{type}{character vector of length one having a column type}
}
\description{
type can be one of c("bead", "syntactic_bead", "summary",
"syntactic_summary")
}
