% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{get_combined}
\alias{get_combined}
\title{Returns dataframe with combined results from antigen results}
\usage{
get_combined(conn, r_mfi_df, file)
}
\arguments{
\item{conn}{database connection object}

\item{r_mfi_df}{antigen results}

\item{file}{character vector with name of file}
}
\description{
The following results set are possible:
 antigen 1 P and antigen 2 P, combined is P
 antigen 1 P and antigen 2 N, combined is I
 antigen 1 N and antigen 2 P, combined in I
 antigen 1 N and antigen 2 N, combined in N
}
