% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_date_from_sample.R
\name{get_date_from_sample}
\alias{get_date_from_sample}
\title{Returns the date from a sample column if it exist else NA.}
\usage{
get_date_from_sample(conn, samples)
}
\arguments{
\item{conn}{database connection object}

\item{samples}{character vector with one or more sample dates}
}
\description{
Returns the date from a sample column if it exist else NA.
}
