% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{get_divisor_mismatch}
\alias{get_divisor_mismatch}
\title{Returns a charcter vector with the file names of files that have region
names that do not
match those found in ``cutoff.xlsx''}
\usage{
get_divisor_mismatch(file_names)
}
\arguments{
\item{file_names}{character vector of file names with XML data}
}
\description{
Returns a charcter vector with the file names of files that have region
names that do not
match those found in ``cutoff.xlsx''
}
