% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_divisors.R
\name{get_divisors}
\alias{get_divisors}
\title{Returns a named numeric vector with the divisors originally provided by the
vendor and copied into the Excel sheet "../inst/extdata/cutoff.xlsx"}
\usage{
get_divisors(file)
}
\arguments{
\item{file}{full or relative path name of the Luminex file being analyzed.}
}
\description{
Returns a named numeric vector with the divisors originally provided by the
vendor and copied into the Excel sheet "../inst/extdata/cutoff.xlsx"
}
