% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{get_excel_divisors}
\alias{get_excel_divisors}
\title{Returns a named numeric vector with the divisors originally provided by the
vendor and copied into the Excel sheet "../inst/extdata/cutoff.xlsx"}
\usage{
get_excel_divisors(results)
}
\arguments{
\item{results}{character vector with results section of the Luminex file
being analyzed.}
}
\description{
Returns a named numeric vector with the divisors originally provided by the
vendor and copied into the Excel sheet "../inst/extdata/cutoff.xlsx"
}
