% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_ids.R
\name{get_file_ids}
\alias{get_file_ids}
\title{Returns integer vector with plate numbers}
\usage{
get_file_ids(files)
}
\arguments{
\item{files}{name of file}
}
\description{
Returns integer vector with plate numbers
}
