% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_followup.R
\name{get_followup}
\alias{get_followup}
\title{Returns ``followup'' assay results, which are groups of assay results for an
animal beginning with the first non-negative result and all subsequent
results.}
\usage{
get_followup(dsn, start_date = "1-1-2013")
}
\arguments{
\item{dsn}{character with the dataset name}

\item{start_date}{character with start date in mm-dd-yyyy format.}
}
\description{
Returns ``followup'' assay results, which are groups of assay results for an
animal beginning with the first non-negative result and all subsequent
results.
}
