% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_followup_pool_samples.R
\name{get_followup_pool_samples}
\alias{get_followup_pool_samples}
\title{Returns assay results for individual animals based on assay results of pooled
samples.}
\usage{
get_followup_pool_samples(followup_df)
}
\arguments{
\item{followup_df}{dataframe with animal Ids to use in selected pool sample
results.}
}
\description{
Returns assay results for individual animals based on assay results of pooled
samples.
}
