% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{get_id_date_repeated}
\alias{get_id_date_repeated}
\title{Returns dataframe with true animal Ids (Id) and date strings (Date) from
Animal.ID column.}
\usage{
get_id_date_repeated(animal_id_col)
}
\arguments{
\item{animal_id_col}{character vector containing the composite descriptor
that has the animal Id and sample date. I may also have an 'R' indicating
that the sample is a repeated sample.}
}
\description{
Returns dataframe with true animal Ids (Id) and date strings (Date) from
Animal.ID column.
}
