% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{get_id_from_sample}
\alias{get_id_from_sample}
\title{Returns the first integer value from a sample column if it exist else NA.}
\usage{
get_id_from_sample(conn, samples)
}
\arguments{
\item{conn}{database connection object}

\item{samples}{character vector with one or more sample Ids}
}
\description{
Returns the first integer value from a sample column if it exist else NA.
}
