% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{get_luminex_species}
\alias{get_luminex_species}
\title{Returns species desination of "CYNO", "MACAQUE" or "PAPIO" at this time.
If the Id does not belong to a cyno, baboon, or rhesus, the arc_species_code
is used.}
\usage{
get_luminex_species(conn, snprc_id)
}
\arguments{
\item{conn}{database connection object}

\item{snprc_id}{character vector with possible animal Ids}
}
\description{
Returns species desination of "CYNO", "MACAQUE" or "PAPIO" at this time.
If the Id does not belong to a cyno, baboon, or rhesus, the arc_species_code
is used.
}
