% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{get_Machine_info}
\alias{get_Machine_info}
\title{Returns Machine serialNumber, name, sftwareVersion, firmwareVersion, and
DSPVersion}
\usage{
get_Machine_info(file)
}
\arguments{
\item{file}{name of file containing XML}
}
\description{
Returns Machine serialNumber, name, sftwareVersion, firmwareVersion, and
DSPVersion
}
