% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{get_mean_raw_mfi}
\alias{get_mean_raw_mfi}
\title{Returns dataframe with means of replicated samples}
\usage{
get_mean_raw_mfi(raw_mfi_df)
}
\arguments{
\item{raw_mfi_df}{dataframe with raw mfi reading}
}
\description{
Returns dataframe with means of replicated samples
}
