% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean_raw_mfi_from_lxd.R
\name{get_mean_raw_mfi_from_lxd}
\alias{get_mean_raw_mfi_from_lxd}
\title{Returns the mean_raw_mfi_df from an XML file (.lxd extension) from the
Luminex machine.}
\usage{
get_mean_raw_mfi_from_lxd(conn, file)
}
\arguments{
\item{conn}{database connection object}

\item{file}{fully qualified file name of XML file}
}
\description{
Returns the mean_raw_mfi_df from an XML file (.lxd extension) from the
Luminex machine.
}
