% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{get_melted_bead_2_wide}
\alias{get_melted_bead_2_wide}
\title{Returns bead based wide dataframe from bead based melted dataframe.}
\usage{
get_melted_bead_2_wide(.df)
}
\arguments{
\item{.df}{bead based melted dataframe}
}
\description{
Returns bead based wide dataframe from bead based melted dataframe.
}
