% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_most_recent_file.R
\name{get_most_recent_file}
\alias{get_most_recent_file}
\title{Returns the most recent assay file name for each plate number}
\usage{
get_most_recent_file(file_id, files_df)
}
\arguments{
\item{file_id}{integer value of plate number}

\item{files_df}{dataframe with file_id, file name and file date.}
}
\description{
Returns the most recent assay file name for each plate number
}
