% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_most_recent_files.R
\name{get_most_recent_files}
\alias{get_most_recent_files}
\title{Returns the most recent assay file name for each plate number}
\usage{
get_most_recent_files(files_all)
}
\arguments{
\item{files_all}{character vector of all file names.}
}
\description{
Returns the most recent assay file name for each plate number
}
