% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pdl_csv_data.R
\name{get_pdl_csv_data}
\alias{get_pdl_csv_data}
\title{Returns pdl data from their CSV files given a list of filenames and the
relative or absolute path.}
\usage{
get_pdl_csv_data(file_names, path)
}
\arguments{
\item{file_names}{character vector with the files to be read}

\item{path}{character string with the path}
}
\description{
Returns pdl data from their CSV files given a list of filenames and the
relative or absolute path.
}
