% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{get_pdl_meta_data_from_content}
\alias{get_pdl_meta_data_from_content}
\title{Returns specified value from a pdl_report$content object}
\usage{
get_pdl_meta_data_from_content(pdl_report, meta_data_label, len)
}
\arguments{
\item{pdl_report}{list containing $content and $meta that make up the
complete data from PDF PDL report file.}

\item{meta_data_label}{character vector of string to search for that
is immediately followed by the meta data being sought}

\item{len}{maximum length of the meta data being retrieved.}
}
\description{
Returns specified value from a pdl_report$content object
}
