% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pdl_nonnegative_reports.R
\name{get_pdl_nonnegative_reports}
\alias{get_pdl_nonnegative_reports}
\title{Returns non-negative results from PDL reports results}
\usage{
get_pdl_nonnegative_reports(pdl_df)
}
\arguments{
\item{pdl_df}{dataframe with data from one or more PDF PDL report files.}
}
\description{
Returns non-negative results from PDL reports results
}
