% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pdl_report_date.R
\name{get_pdl_report_date}
\alias{get_pdl_report_date}
\title{Returns report date from a pdl_report object}
\usage{
get_pdl_report_date(pdl_report)
}
\arguments{
\item{pdl_report}{list containing $content and $meta that make up the
complete data from PDF PDL report file.}
}
\description{
Returns report date from a pdl_report object
}
