% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{get_pdl_report_file_name}
\alias{get_pdl_report_file_name}
\title{Returns report file name from a pdl_report object}
\usage{
get_pdl_report_file_name(pdl_report)
}
\arguments{
\item{pdl_report}{list containing $content and $meta that make up the
complete data from PDF PDL report file.}
}
\description{
Returns report file name from a pdl_report object
}
