% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{get_pdl_reports}
\alias{get_pdl_reports}
\title{Returns a list of lists of length two containing $content and $meta lists
derived from each PDL report.}
\usage{
get_pdl_reports(
  base_directory = "../inst/extdata/",
  pdl_report_directory = "PDL/",
  start_year,
  end_year,
  excluded_files = character(0)
)
}
\arguments{
\item{base_directory}{character vector with path for directory in which the
pdl_report_directory resides.}

\item{pdl_report_directory}{character vector with name of directory in which
the PDL reports reside.}

\item{start_year}{integer value indicating the first calendar year as used in
the name of the folders containing reports from the years indicated.}

\item{end_year}{integer value indicating the last calendar year as used in
the name of the folders containing reports from the years indicated.}

\item{excluded_files}{character vector of files to ignore.}
}
\description{
The returned list contains one list of length 2 for each PDL report.
The 2 elements within each list are the content list, which has one element
for each line within the report and the meta list, which has 7 elements.
See get_pdl_meta_data_from_content().
}
