% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{get_pdl_results}
\alias{get_pdl_results}
\title{Get PDL results}
\usage{
get_pdl_results(conn)
}
\arguments{
\item{conn}{database connection object}
}
\value{
Dataframe with all contents of the pdl_results table except the
timestamp column.
}
\description{
Get PDL results
}
