% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{get_pdl_results_test_comment}
\alias{get_pdl_results_test_comment}
\title{Returns a character vector of length 2 with the result and comment}
\usage{
get_pdl_results_test_comment(line_1, line_2, col_boundaries)
}
\arguments{
\item{line_1}{first line of data}

\item{line_2}{second line of data}

\item{col_boundaries}{locations of start and end column for each data type}
}
\description{
Returns a character vector of length 2 with the result and comment
}
