% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{get_plateID}
\alias{get_plateID}
\title{Returns Plate[@plateID] (plateID attribute of the Plate node.)}
\usage{
get_plateID(lum)
}
\arguments{
\item{lum}{XML object}
}
\description{
Returns Plate[@plateID] (plateID attribute of the Plate node.)
}
