% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{get_pooled_sample_df}
\alias{get_pooled_sample_df}
\title{Returns dataframe with animal Ids, cage location, bleed date, whether not a
blood sample was expected, whether or not a sample was received, and the
Pool Id number.}
\usage{
get_pooled_sample_df(path)
}
\arguments{
\item{path}{charcter vector of length 1 having the path section of the
fully qualified file name.}
}
\description{
Returns dataframe with animal Ids, cage location, bleed date, whether not a
blood sample was expected, whether or not a sample was received, and the
Pool Id number.
}
