% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{get_positive_antigens}
\alias{get_positive_antigens}
\title{Returns  positive antigens based on low control value and a multiplier}
\usage{
get_positive_antigens(d_mfi_df, mult = 1)
}
\arguments{
\item{d_mfi_df}{dataframe wih standardized antigen reactivity measures}

\item{mult}{multiplier used to adjust cutoff level.}
}
\description{
From e-mail 2015-09-22 13:32 by Luis Giavedoni
}
\details{
Mark,

When the CNT LOW is below 2.5, we use the CNT LOW score to assign
positive reactivity. Any score equal or higher than the CNT LOW will be
scored as positive.

When the CNT LOW is above 2.5, we use 2.5 as the cut off. Any score equal
or higher than 2.5 will be scored as positive.

I know that I said that the low score was 3.0, but after review of
current Charles River documents, the recommendation is for the score to be
2.5.

Regards,

Luis

Thus, the function used is
df$result <- ifelse(df$val >= min(2.5, df$cutoff), "P", "N")

\code{SPF Colony Testing Plate 96   06-09-16.xlsx} has an extra low control
\code{ICL CNT LOW (dup)}.
Controls with \code{stri_detect_fixed(tolower(name), "(dup)") == TRUE}
 are ignored.
}
