% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_possible_ids_dates.R
\name{get_possible_ids_dates}
\alias{get_possible_ids_dates}
\title{Returns a list of length 2 with possible$ids and possible$dates}
\usage{
get_possible_ids_dates(samples)
}
\arguments{
\item{samples}{character vector of sample descriptors}
}
\description{
Returns a list of length 2 with possible$ids and possible$dates
}
