% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_mfi.R
\name{get_raw_mfi}
\alias{get_raw_mfi}
\title{Returns dataframe of original values}
\usage{
get_raw_mfi(conn, file)
}
\arguments{
\item{conn}{database connection object}

\item{file}{character vector with filename}
}
\description{
Returns dataframe of original values
}
