% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{get_raw_mfi_df}
\alias{get_raw_mfi_df}
\title{Returns dataframe with raw mfi values from the raw section of the Excel
file.}
\usage{
get_raw_mfi_df(conn, df, file, run_props, run_error)
}
\arguments{
\item{conn}{database connection object}

\item{df}{character vector with line from the Excel file worksheet
having only the result table records}

\item{file}{character vector of length one with filename}

\item{run_props}{dataframe with four columns ("name", "value",
"java_data_type", and "location" defined when a Run Properties File provided
by the system is read by the \code{readRunPropertiesFile()} function.}

\item{run_error}{list of length 2 having the integer value of the error
level and the message to be displayed.}
}
\value{
\code{raw_mfi_df} dataframe
}
\description{
Returns dataframe with raw mfi values from the raw section of the Excel
file.
}
