% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_mfi_tables.R
\name{get_raw_mfi_tables}
\alias{get_raw_mfi_tables}
\title{Returns character vector with the raw data (raw mfi) table}
\usage{
get_raw_mfi_tables(content)
}
\arguments{
\item{content}{character vector with lines from the Excel file worksheet.}
}
\description{
Returns character vector with the raw data (raw mfi) table
}
