% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{get_region_df}
\alias{get_region_df}
\title{Returns a dataframe with region order, region name, and region id}
\usage{
get_region_df(file, setup_name = "DefaultSetup")
}
\arguments{
\item{file}{name of file}

\item{setup_name}{value of name attribute of <Setup>}
}
\description{
Returns a dataframe with region order, region name, and region id
}
