% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{get_repeated_from_sample}
\alias{get_repeated_from_sample}
\title{Returns a character vector of 'Y' if the sample indicates it is a
repeat else 'N'}
\usage{
get_repeated_from_sample(sample)
}
\arguments{
\item{sample}{character vector with one or more sample Ids}
}
\description{
Returns a character vector of 'Y' if the sample indicates it is a
repeat else 'N'
}
