% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_report_date_from_filename.R
\name{get_report_date_from_filename}
\alias{get_report_date_from_filename}
\title{Returns report date in yyyy-mm-dd format as derived from the file names.}
\usage{
get_report_date_from_filename(filenames, run_props, run_error)
}
\arguments{
\item{filenames}{character vector with one or more file names. The
file names of the Excel files have embedded report dates.}

\item{run_props}{dataframe with four columns ("name", "value",
"java_data_type", and "location" defined when a Run Properties File provided
by the system is read by the \code{readRunPropertiesFile()} function.}

\item{run_error}{list of length 2 having the integer value of the error
level and the message to be displayed.}
}
\description{
Returns report date in yyyy-mm-dd format as derived from the file names.
}
