% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{get_result_tables}
\alias{get_result_tables}
\title{Returns character vector with the result tables}
\usage{
get_result_tables(content)
}
\arguments{
\item{content}{character vector with lines from the Excel file worksheet.}
}
\description{
Returns character vector with the result tables
}
