% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{get_sample_df}
\alias{get_sample_df}
\title{Returns dataframe with animal Ids, cage location, bleed date, whether not a
blood sample was expected, whether or not a sample was received, and the
Pool Id number if present.}
\usage{
get_sample_df(all_files, conn, run_props, run_error)
}
\arguments{
\item{all_files}{charcter vector of all the files to process.}

\item{conn}{database connection object.}

\item{run_props}{dataframe with four columns ("name", "value",
"java_data_type", and "location" defined when a Run Properties File provided
by the system is read by the \code{readRunPropertiesFile()} function.}

\item{run_error}{list of length 2 having the integer value of the error
level and the message to be displayed.}
}
\description{
Returns dataframe with animal Ids, cage location, bleed date, whether not a
blood sample was expected, whether or not a sample was received, and the
Pool Id number if present.
}
