% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_type_trans.R
\name{get_sample_type_trans}
\alias{get_sample_type_trans}
\title{Returns the corrected sample type given slightly malformed or incomplete
sample types.}
\usage{
get_sample_type_trans(sample_type)
}
\arguments{
\item{sample_type}{character vector of sample types to be translated.}
}
\description{
Returns the corrected sample type given slightly malformed or incomplete
sample types.
}
