% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_session_data_tm.R
\name{get_session_date_tm}
\alias{get_session_date_tm}
\title{Returns date and time of assay based on '//Plate[@name].}
\usage{
get_session_date_tm(lum)
}
\arguments{
\item{lum}{XML object}
}
\description{
Returns date and time of assay based on '//Plate[@name].
}
