% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_setup_name.R
\name{get_setup_name}
\alias{get_setup_name}
\title{Returns the Setup[@name]s for a lum}
\usage{
get_setup_name(lum, setup = 0)
}
\arguments{
\item{lum}{XML object}

\item{setup}{indicates which Setup to return the name for. 0 indicates
return all setup names.}
}
\description{
Returns the Setup[@name]s for a lum
}
