% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{get_snprc_id}
\alias{get_snprc_id}
\title{Returns an SNPRC animal Id of the pdl_animal_id represents an individual
animal}
\usage{
get_snprc_id(pdl_animal_id)
}
\arguments{
\item{pdl_animal_id}{Id used by PDL to indicate the pooled sample or the
individual animal sample.}
}
\description{
Returns an SNPRC animal Id of the pdl_animal_id represents an individual
animal
}
