% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_start_end_dates.R
\name{get_start_end_dates}
\alias{get_start_end_dates}
\title{Returns data frame with file and start_date as the two columns give
a list of luminex data file names}
\usage{
get_start_end_dates(files)
}
\arguments{
\item{files}{character vector with file names (with relative or absolute
path).}
}
\description{
Values are taken from '//Plate[@startDate]"
}
