% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{get_still_pos}
\alias{get_still_pos}
\title{Get dataframe of animals where the most recent assay was not negative.}
\usage{
get_still_pos(sorted_df)
}
\arguments{
\item{sorted_df}{dataframe sorted values of all assays.}
}
\value{
A dataframe with assays whose last value are non-negative.
}
\description{
A PCR assay trumps an non-PCR assay so that if a PCR assay is negative and
and non-PCR assay is non-negative, the animal is negative for that agent.
}
