% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary_df.R
\name{get_summary_df}
\alias{get_summary_df}
\title{Returns dataframe with summary results from the results section of the Excel
file.}
\usage{
get_summary_df(results)
}
\arguments{
\item{results}{character vector with line from the Excel file worksheet
having only the result table records}
}
\description{
Returns dataframe with summary results from the results section of the Excel
file.
}
