% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{get_test_antigens}
\alias{get_test_antigens}
\title{#' Makes function to return agents and antigen pairs.
#'
#' @param session session object from LabKey.
#' @import Rlabkey
make_get_agents_and_antigen_pairs <- function(session) {
  schema_conn <- getSchema(session, "lists")
  schema_element <- schema_conn[["agent_antigen_pairs"]]
  agents_and_antigen_pairs <- getRows(session, query)
  f <- function() {
    agents_and_antigen_pairs
  }
}}
\usage{
get_test_antigens()
}
\description{
Returns the list of test antigens
}
\details{
Uses \code{get_agents_and_antigen_pairs()} to create the list of unique
test antigens
}
