% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{get_usable_files}
\alias{get_usable_files}
\title{Returns a character vector of names of usable files.}
\usage{
get_usable_files(path)
}
\arguments{
\item{path}{file path to directory containing files to be read.}
}
\description{
Usable files are those from "Report plate 55 03-21-2014.xlsx" forward
because of large fluxuations in format.
}
