% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_version.R
\name{get_version}
\alias{get_version}
\title{getVersion Get the version number of snprcspf}
\usage{
get_version()
}
\value{
Current Version
}
\description{
getVersion Get the version number of snprcspf
}
\examples{
\donttest{
library(snprcspf)
get_version()
}
}
