% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{get_wide_df}
\alias{get_wide_df}
\title{Returns a list of dataframes that have been converted to wide format using
dcast}
\usage{
get_wide_df(mean_raw_mfi_df, mfi_df, d_mfi_df, r_mfi_df, combined_df)
}
\arguments{
\item{mean_raw_mfi_df}{dataframe to be converted to wide format}

\item{mfi_df}{dataframe to be converted to wide format}

\item{d_mfi_df}{dataframe to be converted to wide format}

\item{r_mfi_df}{dataframe to be converted to wide format}

\item{combined_df}{dataframe to be converted to wide format}
}
\description{
Returns a list of dataframes that have been converted to wide format using
dcast
}
