% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_workset_desc.R
\name{get_workset_desc}
\alias{get_workset_desc}
\title{Returns /WorkSet/Description from XML}
\usage{
get_workset_desc(lum)
}
\arguments{
\item{lum}{XML object}
}
\description{
Returns /WorkSet/Description from XML
}
