% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{get_yyyymmdd_from_possible_dates}
\alias{get_yyyymmdd_from_possible_dates}
\title{Returns yyyymmdd in with optional separator between date units if
string looks like a date.}
\usage{
get_yyyymmdd_from_possible_dates(dates, sep = "-")
}
\arguments{
\item{dates}{character vector with possible dates}

\item{sep}{character string to be used to separate date units.}
}
\description{
Returns yyyymmdd in with optional separator between date units if
string looks like a date.
}
