% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRunPropertyValue.R
\name{getRunPropertyValue}
\alias{getRunPropertyValue}
\title{Returns value of run property if one is available else NA is provided.}
\usage{
getRunPropertyValue(run_props, colName)
}
\arguments{
\item{run_props}{dataframe with four columns ("name", "value",
"java_data_type", and "location" defined when a Run Properties File provided
by the system is read by the \code{readRunPropertiesFile()} function.}

\item{colName}{character vector of length 1 with the name of the property
whose value is sought.}
}
\description{
Returns value of run property if one is available else NA is provided.
}
