% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{getTransformedOutputFile}
\alias{getTransformedOutputFile}
\title{Returns full name with path of transformed output file.}
\usage{
getTransformedOutputFile(run_props)
}
\arguments{
\item{run_props}{dataframe with four columns ("name", "value",
"java_data_type", and "location" defined when a Run Properties File provided
by the system is read by the \code{readRunPropertiesFile()} function.}
}
\description{
Returns full name with path of transformed output file.
}
