% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleErrorsAndWarnings.R
\name{handleErrorsAndWarnings}
\alias{handleErrorsAndWarnings}
\title{Returns NULL: does not return -- writes the maximumSeverity level to the
transformRunProperties file and the error message to the error.html file.}
\usage{
handleErrorsAndWarnings(run_props, run_error)
}
\arguments{
\item{run_props}{dataframe with four columns ("name", "value",
"java_data_type", and "location" defined when a Run Properties File provided
by the system is read by the \code{readRunPropertiesFile()} function.}

\item{run_error}{list of length 2 having the integer value of the error
level and the message to be displayed.}
}
\description{
LabKey server will read these files after execution to determine if an error
or warning occurred and handle it appropriately
}
