% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_combined_assay_results.R
\name{insert_combined_assay_results}
\alias{insert_combined_assay_results}
\title{Returns the number of saved records}
\usage{
insert_combined_assay_results(conn, r_df, table = "luminex_screen_results")
}
\arguments{
\item{conn}{database connection object}

\item{r_df}{dataframe containing the melted version of the individual bead}

\item{table}{optional character variable with name of table to insert into.
assay
results.}
}
\description{
Returns the number of saved records
}
