% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{insert_missing_lines}
\alias{insert_missing_lines}
\title{Returns character vector representing the contents of the pdl_report$content
with the missing second lines of data replaced.}
\usage{
insert_missing_lines(content)
}
\arguments{
\item{content}{character vector representing the contents of the
pdl_report$content with the possible missing second lines of data}
}
\description{
Returns character vector representing the contents of the pdl_report$content
with the missing second lines of data replaced.
}
