% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{insert_pdl_results}
\alias{insert_pdl_results}
\title{Returns the number of records inserted into pdl_results}
\usage{
insert_pdl_results(conn, pdl_df)
}
\arguments{
\item{conn}{database connection object}

\item{pdl_df}{dataframe with PDL results}
}
\description{
Returns the number of records inserted into pdl_results
}
