% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_spf_samples.R
\name{insert_spf_samples}
\alias{insert_spf_samples}
\title{Returns the number of records inserted into spf_samples}
\usage{
insert_spf_samples(conn, spf_samples_df, run_props, run_error)
}
\arguments{
\item{conn}{database connection object}

\item{spf_samples_df}{dataframe with samples}

\item{run_props}{dataframe with four columns ("name", "value",
"java_data_type", and "location" defined when a Run Properties File provided
by the system is read by the \code{readRunPropertiesFile()} function.}

\item{run_error}{list of length 2 having the integer value of the error
level and the message to be displayed.}
}
\description{
Returns the number of records inserted into spf_samples
}
