% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_column_labels.R
\name{is_column_labels}
\alias{is_column_labels}
\title{Returns TRUE if a line is an expected set of column labels}
\usage{
is_column_labels(line)
}
\arguments{
\item{line}{character vector of length 1.}
}
\description{
Returns TRUE if a line is an expected set of column labels
}
