% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luminex_agent_to_sqlmed_test_id.R
\name{luminex_agent_to_sqlmed_test_id}
\alias{luminex_agent_to_sqlmed_test_id}
\title{Returns a numeric vector of SqlMed test_id given a character
vector of luminex_agent.}
\usage{
luminex_agent_to_sqlmed_test_id(luminex_agent)
}
\arguments{
\item{luminex_agent}{character vector of pdl_test; case sensitive}
}
\description{
Returns a numeric vector of SqlMed test_id given a character
vector of luminex_agent.
}
