% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luminex_agent_to_sqlmed_test_name.R
\name{luminex_agent_to_sqlmed_test_name}
\alias{luminex_agent_to_sqlmed_test_name}
\title{Returns a character vector of SqlMed test_name given a character
vector of luminex_agent.}
\usage{
luminex_agent_to_sqlmed_test_name(luminex_agent)
}
\arguments{
\item{luminex_agent}{character vector of pdl_test; case sensitive}
}
\description{
Returns a character vector of SqlMed test_name given a character
vector of luminex_agent.
}
