% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luminex_sqlmed_procedure_id.R
\name{luminex_to_sqlmed_procedure_id}
\alias{luminex_to_sqlmed_procedure_id}
\title{Returns a numeric vector of SqlMed procedure_id given a character
vector of luminex_species and luminex_repeatedy.}
\usage{
luminex_to_sqlmed_procedure_id(luminex_species, luminex_repeated)
}
\arguments{
\item{luminex_species}{character vector "MACAQUE" or "PAPIO"}

\item{luminex_repeated}{characer vector with "Y" or "N" depending on
whether the sample ended in "-R" or not.}
}
\description{
Warning luminex does not provide species designations reliably.
This will have to come from post processing.
}
