% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{luminex_to_sqlmed_procedure_name}
\alias{luminex_to_sqlmed_procedure_name}
\title{Returns a character vector of SqlMed procedure_names given a character
vector of Luminex_assay.}
\usage{
luminex_to_sqlmed_procedure_name(luminex_species, luminex_repeated)
}
\arguments{
\item{luminex_species}{character vector of agent; case sensitive}

\item{luminex_repeated}{character vector of whether the assay was
repeated for the animal}
}
\description{
Returns a character vector of SqlMed procedure_names given a character
vector of Luminex_assay.
}
