% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_excel_wkbk.R
\name{make_excel_wkbk}
\alias{make_excel_wkbk}
\title{Form an Excel workbook and worksheet similar in format to those being
made by Luis Giavedoni's laboratory.}
\usage{
make_excel_wkbk(
  file,
  w_raw_mfi_df,
  w_mfi_df,
  w_d_mfi_df,
  w_r_mfi_df,
  w_combined_df,
  low_positive_controls_df
)
}
\arguments{
\item{file}{character vector of length one with name of file to create}

\item{w_raw_mfi_df}{dataframe with raw results from Luminex machine}

\item{w_mfi_df}{dataframe with background substracted from raw results}

\item{w_d_mfi_df}{dataframe with normalized values where divisors were
used on background subtracted results}

\item{w_r_mfi_df}{dataframe where each antigen result is called as N, I, or
P based on is values relative to a fraction of the low control value.}

\item{w_combined_df}{dataframe with the antigen calls combined into agent
(virus) calls. (See get_combined() for algorithm).}

\item{low_positive_controls_df}{dataframe with \code{sample}, \code{name},
\code{wells}, and \code{val} columns that identify the well with low
positive control values (<5000 for Human or Goat IgG).}
}
\description{
The Excel file ``Report plate 62 08-15-2014.xlsx'' was used as a guide.
}
