% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{mdy_to_yyyymmdd}
\alias{mdy_to_yyyymmdd}
\title{Returns character vector of dates where entities of mm/dd/yy or
mm-dd-yy are converted to yyyymmdd.}
\usage{
mdy_to_yyyymmdd(dates, sep = "")
}
\arguments{
\item{dates}{character vector of potential dates}

\item{sep}{character vector of length one having the character(s) to be used
to separate elements with each date.}
}
\value{
date with format of yyyymmdd in character vector
}
\description{
Returns character vector of dates where entities of mm/dd/yy or
mm-dd-yy are converted to yyyymmdd.
}
