% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{merge_pooled_and_pdl}
\alias{merge_pooled_and_pdl}
\title{Returns a dataframe with pooled_df merged with pdl_df by pooled_id with
pdl_animal_id. All pooled records are retained.}
\usage{
merge_pooled_and_pdl(pooled_df, pdl_df)
}
\arguments{
\item{pooled_df}{dataframe with pooled samples}

\item{pdl_df}{dataframe with PDL results.}
}
\description{
Returns a dataframe with pooled_df merged with pdl_df by pooled_id with
pdl_animal_id. All pooled records are retained.
}
