% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_to_char.R
\name{na_to_char}
\alias{na_to_char}
\title{Replaces an NA value with a character string}
\usage{
na_to_char(x, char)
}
\arguments{
\item{x}{character vector}

\item{char}{character string used to replace NAs}
}
\description{
Replaces an NA value with a character string
}
