% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{pdl_assay_to_sqlmed_procedure_name}
\alias{pdl_assay_to_sqlmed_procedure_name}
\title{Returns a character vector of SqlMed procedure_names given a character
vector of pdl_assay.}
\usage{
pdl_assay_to_sqlmed_procedure_name(pdl_assay)
}
\arguments{
\item{pdl_assay}{character vector of pdl_test; case sensitive}
}
\description{
Returns a character vector of SqlMed procedure_names given a character
vector of pdl_assay.
}
