% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdl_assay_to_sqlmed_test_id.R
\name{pdl_assay_to_sqlmed_test_id}
\alias{pdl_assay_to_sqlmed_test_id}
\title{Returns a character vector of SqlMed test_id given a character
vector of pdl_assay.}
\usage{
pdl_assay_to_sqlmed_test_id(pdl_assay)
}
\arguments{
\item{pdl_assay}{character vector of pdl_test; case sensitive}
}
\description{
Returns a character vector of SqlMed test_id given a character
vector of pdl_assay.
}
