% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdl_report_td_df.R
\name{pdl_report_to_df}
\alias{pdl_report_to_df}
\title{Returns dataframe with parsed content of pdl_report}
\usage{
pdl_report_to_df(pdl_report)
}
\arguments{
\item{pdl_report}{list containing $content and $meta that make up the
complete data from PDF PDL report file.}
}
\description{
Returns dataframe with parsed content of pdl_report
}
