% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{pdl_reports_to_df}
\alias{pdl_reports_to_df}
\title{Returns dataframe with results from all report files in directory.}
\usage{
pdl_reports_to_df(base_directory, pdl_report_directory, start_year,
  end_year, excluded_files = character(0))
}
\arguments{
\item{base_directory}{path of base directory
base_directory = ".."}

\item{pdl_report_directory}{name of directory containing PDL reports
pdl_report_directory = "inst/extdata/"}

\item{start_year}{integer value indicating the first calendar year as used in
the name of the folders containing reports from the years indicated.}

\item{end_year}{integer value indicating the last calendar year as used in
the name of the folders containing reports from the years indicated.}

\item{excluded_files}{character vector of files within the PDL directory
that are to be ignored.
excluded_files = character(0))}
}
\description{
Returns dataframe with results from all report files in directory.
}
