% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plate165_dfs}
\alias{plate165_dfs}
\title{plate165_dfs is a loadable version of the dfs list of dataframes
created by `get_wide_df`
used by `make_excel_wkbk`.}
\format{
An object of class \code{list} of length 5.
}
\usage{
plate165_dfs
}
\description{
plate165_dfs is a loadable version of the dfs list of dataframes
created by `get_wide_df`
used by `make_excel_wkbk`.
}
\examples{
\dontrun{
dfs <- snprcspf::plate165_dfs
low_positive_controls_df <- data.frame()
make_excel_wkbk(excel_file, dfs$w_mean_raw_mfi_df, dfs$w_mfi_df,
                dfs$w_d_mfi_df, dfs$w_r_mfi_df, dfs$w_combined_df,
                low_positive_controls_df = low_positive_controls_df)
}
}
\keyword{datasets}
