% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plate165_mean_raw_mfi_df}
\alias{plate165_mean_raw_mfi_df}
\title{plate165_mean_raw_mfi_df is a loadable version of the mean_raw_mfi_df dataframe
created by `get_mean_raw_mfi_from_xlsx` or `get_mean_raw_mfi_from_lxd` and
used by `get_low_positive_controls` and `subtract_background`.}
\format{An object of class \code{data.frame} with 1092 rows and 8 columns.}
\usage{
plate165_mean_raw_mfi_df
}
\description{
plate165_mean_raw_mfi_df is a loadable version of the mean_raw_mfi_df dataframe
created by `get_mean_raw_mfi_from_xlsx` or `get_mean_raw_mfi_from_lxd` and
used by `get_low_positive_controls` and `subtract_background`.
}
\examples{
\dontrun{
plate165_mean_raw_mfi_df <- snprcspf::plate165_mean_raw_mfi_df
}
}
\keyword{datasets}
