% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plate166_combined_df}
\alias{plate166_combined_df}
\title{combined_df is a loadable version of the combined_df dataframe
created by `get_combined`
used by `get_wide_df`, `print_bad_animal_ids` and `print_bad_sample_dates`.}
\format{An object of class \code{data.frame} with 552 rows and 17 columns.}
\usage{
plate166_combined_df
}
\description{
combined_df is a loadable version of the combined_df dataframe
created by `get_combined`
used by `get_wide_df`, `print_bad_animal_ids` and `print_bad_sample_dates`.
}
\examples{
\dontrun{
plate166_combined_df <- snprcspf::plate166_combined_df
}
}
\keyword{datasets}
