% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plate166_r_mfi_df}
\alias{plate166_r_mfi_df}
\title{plate166_r_mfi_df is a loadable version of the r_mfi_df dataframe
created by `r_mfi_df` followed by `basename`
used by `basename` (with `basename(r_mfi_df$file)`, `get_combined` and
`get_wide_df`.}
\format{An object of class \code{data.frame} with 960 rows and 10 columns.}
\usage{
plate166_r_mfi_df
}
\description{
plate166_r_mfi_df is a loadable version of the r_mfi_df dataframe
created by `r_mfi_df` followed by `basename`
used by `basename` (with `basename(r_mfi_df$file)`, `get_combined` and
`get_wide_df`.
}
\examples{
\dontrun{
plate166_r_mfi_df <- snprcspf::plate166_r_mfi_df
basename(plate166_r_mfi_df$file)
}
}
\keyword{datasets}
