% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooled_bleed_date_to_sample_date.R
\name{pooled_bleed_date_to_sample_date}
\alias{pooled_bleed_date_to_sample_date}
\title{Returns pooled_df dataframe with bleed_date replaced by sample_date.}
\usage{
pooled_bleed_date_to_sample_date(pooled_df, sep = "-")
}
\arguments{
\item{pooled_df}{dataframe with bleed_date in yyyymmdd format.}

\item{sep}{character vector of length one having the character(s) to be used}
}
\value{
\code{pooled_df} dataframe with \code{sample_date} in yyyy-mm-dd
format.
}
\description{
Returns pooled_df dataframe with bleed_date replaced by sample_date.
}
