% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{pooled_Y_or_N}
\alias{pooled_Y_or_N}
\title{Returns "Y" if a sample is pooled and "N" if it is an individual sample.}
\usage{
pooled_Y_or_N(pdl_animal_id)
}
\arguments{
\item{pdl_animal_id}{Id used by PDL to indicate the pooled sample or the
individual animal sample.}
}
\description{
Returns "Y" if a sample is pooled and "N" if it is an individual sample.
}
