% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{print_bad_animal_ids}
\alias{print_bad_animal_ids}
\title{Returns a dataframe of file names, sample, snprc_id, and sample_date and
has the side effect
of printing out the bad Ids in a table with file name LocName}
\usage{
print_bad_animal_ids(combined_df, type = "latex", ...)
}
\arguments{
\item{combined_df}{dataframe with agent assay data results and meta data}

\item{type}{character string indicating if the output is to be in "latex"
format or "html".}

\item{...}{extra arguments to \code{xt_print()} that go to \code{print()}}
}
\description{
Returns a dataframe of file names, sample, snprc_id, and sample_date and
has the side effect
of printing out the bad Ids in a table with file name LocName
}
