% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sample_file.R
\name{read_sample_file}
\alias{read_sample_file}
\title{Read blood or dna sample file}
\usage{
read_sample_file(file_name, conn, run_props, run_error)
}
\arguments{
\item{file_name}{fully qualified file name of Excel file}

\item{conn}{database connection object.}

\item{run_props}{dataframe with four columns ("name", "value",
"java_data_type", and "location" defined when a Run Properties File provided
by the system is read by the \code{readRunPropertiesFile()} function.}

\item{run_error}{list of length 2 having the integer value of the error
level and the message to be displayed.}
}
\value{
dataframe with animal Ids, cage location, bleed date, whether not a
blood sample was expected, whether or not a sample was received, and the
Pool Id number if present for a single sample file.
}
\description{
Read blood or dna sample file
}
