% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRunPropertiesFile.R
\name{readRunPropertiesFile}
\alias{readRunPropertiesFile}
\title{Returns a dataframe with four columns ("name", "value",
"java_data_type", and "location"), which is defined when a Run Properties
File provided by the system is read.}
\usage{
readRunPropertiesFile(properties_file)
}
\arguments{
\item{properties_file}{character string with the file name provided by the
system as a replacement value for the macro \emph{${runInfo}}}
}
\description{
Returns a dataframe with four columns ("name", "value",
"java_data_type", and "location"), which is defined when a Run Properties
File provided by the system is read.
}
