% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_bad_animal_ids.R
\name{remove_bad_animal_ids}
\alias{remove_bad_animal_ids}
\title{Returns only good SNPRC animal Ids given a character vector of potential Ids}
\usage{
remove_bad_animal_ids(conn, ids)
}
\arguments{
\item{conn}{database connection object}

\item{ids}{character vector of potential animal Ids.}
}
\description{
Returns only good SNPRC animal Ids given a character vector of potential Ids
}
