% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{remove_bn_and_dash_R}
\alias{remove_bn_and_dash_R}
\title{Returns samples with "Bn" and "-R" trimmed off of each end}
\usage{
remove_bn_and_dash_R(samples)
}
\arguments{
\item{samples}{character vector of sample descriptors}
}
\description{
Returns samples with "Bn" and "-R" trimmed off of each end
}
