% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{remove_pdl_report_headers}
\alias{remove_pdl_report_headers}
\title{Returns content for PDL reports without introductory material and end of
page material}
\usage{
remove_pdl_report_headers(content)
}
\arguments{
\item{content}{copy of pdl_report$content taken from PDF PDL report}
}
\description{
Returns content for PDL reports without introductory material and end of
page material
}
