% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_bad_csv_files.R
\name{replace_bad_csv_files}
\alias{replace_bad_csv_files}
\title{Returns a dataframe with a row for each bad CSV file with the original
name and the renamed file}
\usage{
replace_bad_csv_files(path, file_names, orig_prefix = "org")
}
\arguments{
\item{path}{character vector of length one with the file path}

\item{file_names}{character vector of the file names to be read}

\item{orig_prefix}{character string with the prefix to be put on the
renamed bad file.}
}
\description{
Some of the files from PDL have text trailing the CSV portion of the file.
This routine removes files from the end until a record that starts with
a token of all integers is found. This is the order primary key.
The bad file has been replaces with a cleaned CSV
file having the original name.
}
