% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_missing_samples.R
\name{report_missing_samples}
\alias{report_missing_samples}
\title{Creates an Excel file with a list of samples that were expected but not
received.}
\usage{
report_missing_samples(sample_df)
}
\arguments{
\item{sample_df}{dataframe from sample file}
}
\description{
This needs testing.
}
