% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_unexpected_samples.R
\name{report_unexpected_samples}
\alias{report_unexpected_samples}
\title{Returns a list of samples that were expected but not
received.}
\usage{
report_unexpected_samples(sample_df, file_name)
}
\arguments{
\item{sample_df}{dataframe from sample file}

\item{file_name}{fully qualified file name of Excel file}
}
\description{
Returns a list of samples that were expected but not
received.
}
