% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_excel_sheet.R
\name{save_excel_sheet}
\alias{save_excel_sheet}
\title{Returns the file name of the Excel file created which contains the
dataframes provided as an arguments to the function.}
\usage{
save_excel_sheet(sorted_df, still_pos_df)
}
\arguments{
\item{sorted_df}{a datafrome containing the sorted followup assay results.}

\item{still_pos_df}{a datafrome containing the sorted followup assay results
by animal where the assay results are still non-negative.}
}
\value{
name of Excel file saved containing the sorted followup dataframe
(\code{sorted_df}).
}
\description{
Returns the file name of the Excel file created which contains the
dataframes provided as an arguments to the function.
}
