% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{setMaxSeverity}
\alias{setMaxSeverity}
\title{Returns integer value of error level (0:NONE, 1:WARN, 2:ERROR) used to set
run_error$level.}
\usage{
setMaxSeverity(run_props, run_error, level)
}
\arguments{
\item{run_props}{dataframe with four columns ("name", "value",
"java_data_type", and "location") defined when a Run Properties File provided}

\item{run_error}{list of length 2 having the integer value of the error
level and the message to be displayed.}

\item{level}{integer value being used to reset run_error$level}
}
\description{
If the new level is 2 then 2 is returned.
If the new level is the same as the current level and not 2, 0 is returned.
If the new level is greater than the current level and not 2 (that is if it
is 1) and the current severity level is not "ERROR", the new level (1) is
returned.
If the new level is 1 and the current severity level is "ERROR", 0 is
returned.
}
