% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_occurances.R
\name{sort_occurances}
\alias{sort_occurances}
\title{Returns a sorted dataframe with First occurrances appearing in
chronological order followed by all subsequent assay results.
Sample results based on pooled samples are included. If an animal is lost
to followup due to leaving the institution or death a record indicating that
event and date is the last record in the group.}
\usage{
sort_occurances(followup_df)
}
\arguments{
\item{followup_df}{dataframe with all records.}
}
\value{
sorted \code{followup_df} as \code{sorted_df}
}
\description{
Returns a sorted dataframe with First occurrances appearing in
chronological order followed by all subsequent assay results.
Sample results based on pooled samples are included. If an animal is lost
to followup due to leaving the institution or death a record indicating that
event and date is the last record in the group.
}
