% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spf_odbcConnect.R
\name{spf_odbcConnect}
\alias{spf_odbcConnect}
\title{Returns database connection object}
\usage{
spf_odbcConnect(spf_dsn)
}
\arguments{
\item{spf_dsn}{DSN for the SPF user}
}
\description{
Returns database connection object
}
