% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{strip_column_labels}
\alias{strip_column_labels}
\title{Returns content minus column labels. Fails with a call to stop()
if anticipated labels are not present}
\usage{
strip_column_labels(content)
}
\arguments{
\item{content}{character vector that has lines of report with data.}
}
\description{
Returns content minus column labels. Fails with a call to stop()
if anticipated labels are not present
}
