% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract_background.R
\name{subtract_background}
\alias{subtract_background}
\title{Returns mfi_df dataframe values with wBAC column subtracted from each
value and the wBAC and the Human IgG, wBAC and Goat anti-human IgG rows
removed from raw_mfi_df}
\usage{
subtract_background(raw_mfi_df)
}
\arguments{
\item{raw_mfi_df}{dataframe with raw data}
}
\description{
Returns mfi_df dataframe values with wBAC column subtracted from each
value and the wBAC and the Human IgG, wBAC and Goat anti-human IgG rows
removed from raw_mfi_df
}
