% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_for_cage.R
\name{test_for_cage}
\alias{test_for_cage}
\title{Throws and error on one or more bad cage numbers are provided.
It terminates the script with an informative
message saying what was wrong otherwise it simply returns NULL.}
\usage{
test_for_cage(cage, file_name, conn, run_props, run_error)
}
\arguments{
\item{cage}{numeric vector of \code{cage}
column of a blood sample file.}

\item{file_name}{character vector of length one with the file name.}

\item{conn}{database connection object to the animal database}

\item{run_props}{dataframe with four columns ("name", "value",
"java_data_type", and "location" defined when a Run Properties File provided
by the system is read by the \code{readRunPropertiesFile()} function.}

\item{run_error}{list of length 2 having the integer value of the error
level and the message to be displayed.}
}
\value{
NULL if test finds no bad input.
}
\description{
Throws and error on one or more bad cage numbers are provided.
It terminates the script with an informative
message saying what was wrong otherwise it simply returns NULL.
}
