% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_for_yes_no.R
\name{test_for_yes_no}
\alias{test_for_yes_no}
\title{Throws and error on bad input for the \code{expected_blood}
or \code{received_blood} column
that terminates the script with an informative
message saying what was wrong otherwise it simply returns NULL.}
\usage{
test_for_yes_no(expected, file_name, col_name, run_props, run_error)
}
\arguments{
\item{expected}{character vector of values from the \code{expected_blood}
or \code{received_blood} column of a blood sample file.}

\item{file_name}{character vector of length one with the file name.}

\item{col_name}{character vector of length one with column name.}

\item{run_props}{dataframe with four columns ("name", "value",
"java_data_type", and "location" defined when a Run Properties File provided
by the system is read by the \code{readRunPropertiesFile()} function.}

\item{run_error}{list of length 2 having the integer value of the error
level and the message to be displayed.}
}
\value{
NULL if test finds no bad input.
}
\description{
Throws and error on bad input for the \code{expected_blood}
or \code{received_blood} column
that terminates the script with an informative
message saying what was wrong otherwise it simply returns NULL.
}
