% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprcspf.R
\name{triggerError}
\alias{triggerError}
\title{Returns NULL: does not return -- triggers error and stops execution}
\usage{
triggerError(run_props, run_error, msg)
}
\arguments{
\item{run_props}{dataframe with four columns ("name", "value",
"java_data_type", and "location" defined when a Run Properties File provided
by the system is read by the \code{readRunPropertiesFile()} function.}

\item{run_error}{list of length 2 having the integer value of the error
level and the message to be displayed.}

\item{msg}{character string containing the warning message}
}
\description{
Returns NULL: does not return -- triggers error and stops execution
}
