% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_luminex_results.R
\name{fmt_luminex_results}
\alias{fmt_luminex_results}
\title{Conditionally formats a worksheet based on data within the dataframe as a
side effect.}
\usage{
fmt_luminex_results(
  excel_file,
  dFrame,
  sheet_name = "final_result",
  low_positive_controls_df
)
}
\arguments{
\item{excel_file}{character vector with the name of the Excel file}

\item{dFrame}{dataframe being used to guide highlighting. It is the same
one as was used to make the worksheet.}

\item{sheet_name}{name of the worksheet to highlight.}

\item{low_positive_controls_df}{dataframe with the low positive controls}
}
\value{
Excel file name of file that has been formatted to highlight
non-negative results in the sheet indicated. This highlights "I" and
"P" results
}
\description{
Conditionally formats a worksheet based on data within the dataframe as a
side effect.
}
