% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_antigens.R
\name{get_test_antigens}
\alias{get_test_antigens}
\title{Returns the list of test antigens}
\usage{
get_test_antigens()
}
\description{
Uses \code{get_agents_and_antigen_pairs()} to create the list of unique
test antigens
}
